moving healthcheck to non api package
