+++ bundle/src/main/java/com/adobe/acs/commons/adobeio/core/service/impl/AdobeioHealthcheck.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package com.adobe.acs.commons.adobeio.core.healthcheck;
+package com.adobe.acs.commons.adobeio.core.service.impl;
 
 import org.apache.sling.hc.api.HealthCheck;
 import org.apache.sling.hc.api.Result;
